Adds permit fns for removing liquidity, code too large.
