+++ test/lib/constants.js
@@ -18,6 +18,7 @@ const ERR_CODES = {
     ERR_TRANSFER_OUT_FAIL: "ERR_TRANSFER_OUT_FAIL",
     ERR_FEED_INVALID: "ERR_FEED_INVALID",
     ERR_ZERO_LIQUIDITY: "ERR_ZERO_LIQUIDITY",
+    ERC20_TRANSFER_AMOUNT: "ERC20: transfer amount exceeds balance",
 };
 
 // CALCULATION PARAMETERS
